+++ src/app/modules/shared/shared.module.ts
@@ -2,9 +2,11 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
 import { MatDividerModule } from '@angular/material/divider';
+import { MatTabsModule } from '@angular/material/tabs';
 import { FreelingDisplayComponent } from './components/freeling-display/freeling-display.component';
 import { GraphvizDisplayComponent } from './components/graphviz-display/graphviz-display.component';
 import { LabeledDataComponent } from './components/labeled-data/labeled-data.component';
+import { SimilarityMatricesDisplayComponent } from './components/similarity-matrices-display/similarity-matrices-display.component';
 import { TextStatusComponent } from './components/text-status/text-status.component';
 import { DepNodeToDotPipe } from './pipes/dep-node-to-dot.pipe';
 
@@ -14,17 +16,20 @@ import { DepNodeToDotPipe } from './pipes/dep-node-to-dot.pipe';
 		TextStatusComponent,
 		FreelingDisplayComponent,
 		GraphvizDisplayComponent,
-  DepNodeToDotPipe
+		DepNodeToDotPipe,
+		SimilarityMatricesDisplayComponent
 	],
 	exports: [
 		LabeledDataComponent,
 		TextStatusComponent,
-		FreelingDisplayComponent
+		FreelingDisplayComponent,
+		SimilarityMatricesDisplayComponent
 	],
 	imports: [
 		CommonModule,
 		MatCardModule,
-		MatDividerModule
+		MatDividerModule,
+		MatTabsModule
 	]
 })
 export class SharedModule {}