+++ src/app/modules/texts/dtos/get-text-summarized.dto.ts
@@ -1,0 +1,5 @@
+import { GetTextBaseDTO } from './get-text-base.dto';
+
+export interface GetTextSummarizedDTO extends GetTextBaseDTO {
+
+}