+++ CHANGELOG.md
@@ -9,7 +9,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Added single student answer view.
+
 ### Changed
+
+- Small navigation fixes
 
 ### Deprecated